+++ src/common/jwt.common.module.ts
@@ -1,0 +1,19 @@
+import { Module } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import { JwtModule } from '@nestjs/jwt';
+
+import { RedisModule } from 'src/redis/redis.module';
+
+@Module({
+  imports: [
+    JwtModule.registerAsync({
+      useFactory: (config: ConfigService) => ({
+        secret: config.get<string>('JWT_SECRET_KEY'),
+      }),
+      inject: [ConfigService],
+    }),
+    RedisModule,
+  ],
+  exports: [JwtModule, RedisModule],
+})
+export class JwtCommonModule {}
