+++ src/user/user.module.ts
@@ -1,26 +1,15 @@
 import { Module } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { JwtModule } from '@nestjs/jwt';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { JwtAuthGuard } from 'src/auth/jwt.guard';
+import { JwtAuthGuard } from 'src/guards/jwt.guard';
 
-import { RedisModule } from 'src/redis/redis.module';
+import { JwtCommonModule } from 'src/common/jwt.common.module';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { User } from './entities/user.entity';
 
 @Module({
-  imports: [
-    JwtModule.registerAsync({
-      useFactory: (config: ConfigService) => ({
-        secret: config.get<string>('JWT_SECRET_KEY'),
-      }),
-      inject: [ConfigService],
-    }),
-    TypeOrmModule.forFeature([User]),
-    RedisModule,
-  ],
+  imports: [JwtCommonModule, TypeOrmModule.forFeature([User])],
   providers: [UserService, JwtAuthGuard],
   controllers: [UserController],
   exports: [UserService],
